summarizer/requirements.txt
google-generativeai
uvicorn
fastapi
python-dotenv
uvicorn
